feat(mock-auth): refactor login/signup page
