+++ src/features/auth/validation/signinSchema.ts
@@ -1,0 +1,6 @@
+import { object, string } from 'yup'
+
+export const signinSchema = object({
+  email: string().required('Email is required').email('Invalid email address'),
+  password: string().required('Password is required'),
+})
